Step 5: wire the client side
The bootstrap quartet (CG_VR_Init / RegisterMedia / Frame / Shutdown),
the whole-frame fork in CG_DrawActive, the CG_VR_AdjustFrom640 keystone,
the per-frame weapon-adjust step (after CG_ProcessSnapshots, never inside
CG_VR_Frame) with its four console commands in the command table, and
the head-tracking protocol hooks in CG_PlayerAngles / CG_CalcViewValues /
CG_InterpolatePlayerState. The five shapes of one-liner call-outs:
view/weapon/embodiment gates, suppression predicates ANDed into existing
draw conditions, event hooks at their stock event sites, the voting and
scoreboard-cursor accessor families (CG_ProcessVoteHold owns the
hold-to-vote timer and command dispatch), and the drop-state resets.
CG_GetProjectionCenter and CG_Draw2DMinimal upgrade to their vr-reading
forms now that the drop links.

Placements the headset proves out: CG_Draw3DModel sets refdef.isHUD on
its mini-scene refdef (scoreboard heads, status-bar face, 3D icons onto
the HUD surface); cgs.media.friendShader registers in every gametype
(the wheel's selection marker draws through it); CG_DrawCrosshair3D
takes the weapon-aimed body with explicit sprite color and a guarded
stereo-separation trace. Sprites need no cgame orientation help - the
engine owns billboarding (score plums stay stock; the crosshair draws
with no counter-roll).

cgs.cursorX/cursorY become float (the drop's CG_VR_ScoreboardCursor
writes the cursor through float *) - the cgame twin of the q3_ui float
cursor fields.

Trinity-only features stayed behind: TV scrub, damage-border vignette,
kick angles, player shadows, the clickable scoreboard's activation side,
and the zoom-scope reticle that lives in the intentionally-empty
CG_DrawScreen2D. Sentinel now survives in qagame and cgame for both
configs; the UI QVMs wake in Step 6.

+++ code/cgame/cg_local.h
@@ -1053,8 +1053,8 @@ typedef struct {
 	int				teamChatPos;
 	int				teamLastChatPos;
 
-	int cursorX;
-	int cursorY;
+	float cursorX;
+	float cursorY;
 	qboolean eventHandling;
 	qboolean mouseCaptured;
 	qboolean sizingHud;
